ClinIQLink_Sample-dataset/sample_submission/requirements.txt
#Using python 3.10! 
numpy<2
scikit-learn
sentence-transformers
nltk
rouge-score
transformers # only required if using GPT-2 example
torch
faiss-cpu
datasets
